Core/Misc: Rename AzeriteItem packet namespace to just Azerite
