+++ src/server/game/Server/Packets/AllPackets.h
@@ -23,7 +23,7 @@
 #include "ArtifactPackets.h"
 #include "AuctionHousePackets.h"
 #include "AuthenticationPackets.h"
-#include "AzeriteItemPackets.h"
+#include "AzeritePackets.h"
 #include "BankPackets.h"
 #include "BlackMarketPackets.h"
 #include "BattlefieldPackets.h"
